Use `parse` consistently in specs
Usage is still a little awkward though. `parse` now works in any
`MonadFail` which `it` blocks are, but `describe` blocks aren't.
So we have a split of direct `parse`s in some specs, but have also
introduced a `parseRight` to better support the pattern of parsing
at the top of a describe block before making several assertions.

We'll want to clean this up further at some point in the future, but
at least this ensures that we're using the same parsing logic (with
the new end-of-input checking) in dev and test.